requirements.txt
flask
databricks-sql-connector
pandas